🚧 Implementing state products
