+++ src/app/app.config.ts
@@ -10,6 +10,8 @@ import { ErrorInterceptor } from './core/interceptors/errors/errors.interceptor'
 import { provideStore } from '@ngrx/store';
 import { provideEffects } from '@ngrx/effects';
 
+import { productsReducer } from './state/products.reducer';
+
 export const appConfig: ApplicationConfig = {
   providers: [
     provideRouter(routes),
@@ -18,7 +20,9 @@ export const appConfig: ApplicationConfig = {
     { provide: HTTP_INTERCEPTORS, useClass: LoggingInterceptor, multi: true },
     { provide: HTTP_INTERCEPTORS, useClass: HeadersInterceptor, multi: true },
     { provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true },
-    provideStore(),
+    provideStore({
+      products: productsReducer
+    }),
     provideEffects()
-]
+  ]
 };
