+++ src/app/state/products.actions.ts
@@ -1,0 +1,12 @@
+import { createAction, props } from "@ngrx/store";
+import { IProduct } from "../features/products/products-list/products.model";
+
+export const loadProducts = createAction(
+    '[Products] Load Products',
+    props<{ products: IProduct[] }>(),
+);
+
+export const loadProductsFailure = createAction(
+    '[Products] Load Products Failure',
+    props<{ error: any }>(),
+)
